working helpers to build the request
